patient_service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=3.13.0
mysqlclient>=2.2.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
python-decouple>=3.8
